backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
SQLAlchemy==2.0.12
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.5.4
pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.7.0.72
librosa==0.10.0.post2
scikit-learn==1.2.2
scipy==1.10.1
transformers==4.28.1